+++ src/library/pkgcache/R/packages-gz.R
@@ -308,8 +308,13 @@ packages_make_sources <- function(mirror, platform, target, repodir,
   # We don't use Archive for recommended packages, until this is
   # fixed in RSPM: https://github.com/rstudio/package-manager/issues/10471
   # To work around: https://github.com/r-lib/pak/issues/467
+
+  # We also don't use it on PPM (Linux), because PPM might serve a source
+  # package on Archive, or a different binary build:
+  # https://github.com/r-lib/pak/issues/623
   recommended <- package %in% recommended_packages()
-  cransrc <- type == "cran" & platform == "source" & !recommended
+  ppm <- grepl("__linux__", url)
+  cransrc <- type == "cran" & platform == "source" & !recommended & !ppm
   result[cransrc] <- zip_vecs(url[cransrc], url2[cransrc])
 
   others <- vlapply(result, is.null)