+++ src/library/pkgcache/NEWS.md
@@ -1,3 +1,12 @@
+# pkgcache (development version)
+
+* The metadata cache now does not use source URLs for packages in `Archive`
+  on Posit Package Manager repositories. This URLs may serve a different
+  package, even a source package when the main URL for the same package
+  serves a binary package. The alternative URLs are not needed on PPM,
+  anyway, because PPM is in a consistent state w.r.t. metadata and
+  package files (https://github.com/r-lib/pak/issues/623).
+
 # pkgcache 2.2.2
 
 * pkgcache now treats R 4.5.x (current R-devel) macOS binaries